decimal -> double fixes + small .gitignore update :)

+++ AI_NETCORE_API/Domain/Creators/BuyOffer/Request/Abstract/IBuyOfferCreateRequest.cs
@@ -8,6 +8,6 @@ namespace Domain.Creators.BuyOffer.Request.Abstract
     {
         int CompanyId { get; }
         int Amount { get; }
-        decimal Price { get; }
+        double Price { get; }
     }
 }
